cosmetics

+++ modules/gallery/helpers/auth.php
@@ -1,4 +1,5 @@
 <?php defined('SYSPATH') || die('No direct script access.');
+
 /**
  * Gallery - a web based photo album viewer and editor
  * Copyright (C) 2000-2013 Bharat Mediratta
@@ -25,12 +26,7 @@ class auth_Core
         $form->set_attr('class', 'g-narrow');
         $form->hidden('continue_url')->value(Session::instance()->get('continue_url'));
         $group = $form->group('login')->label(t('Login'));
-        $group->input('name')->label(t('Username'))->id('g-username')->class(null)
-              ->callback('auth::validate_too_many_failed_logins')
-              ->error_messages(
-                  'too_many_failed_logins',
-                  t('Too many failed login attempts.  Try again later')
-      );
+        $group->input('name')->label(t('Username'))->id('g-username')->class(null)->callback('auth::validate_too_many_failed_logins')->error_messages('too_many_failed_logins', t('Too many failed login attempts.  Try again later'));
         $group->password('password')->label(t('Password'))->id('g-password')->class(null);
         $group->inputs['name']->error_messages('invalid_login', t('Invalid name or password'));
         $group->submit('')->value(t('Login'));
@@ -42,7 +38,7 @@ class auth_Core
         identity::set_active_user($user);
         if (identity::is_writable()) {
             $user->login_count += 1;
-            $user->last_login = time();
+            $user->last_login  = time();
             $user->save();
         }
         log::info('user', t('User %name logged in', ['name' => $user->name]));
@@ -60,17 +56,10 @@ class auth_Core
             }
             module::event('user_logout', $user);
         }
-        log::info(
-            'user',
-            t('User %name logged out', ['name' => $user->name]),
-            t(
-                '<a href="%url">%user_name</a>',
-                [
-                    'url'       => user_profile::url($user->id),
-                    'user_name' => html::clean($user->name)
-                ]
-              )
-    );
+        log::info('user', t('User %name logged out', ['name' => $user->name]), t('<a href="%url">%user_name</a>', [
+                                                                                                                    'url'       => user_profile::url($user->id),
+                                                                                                                    'user_name' => html::clean($user->name)
+                                                                                                                ]));
     }
 
     /**
@@ -79,12 +68,10 @@ class auth_Core
      */
     public static function too_many_failures($name)
     {
-        $failed = ORM::factory('failed_auth')
-      ->where('name', '=', $name)
-      ->find();
-        return ($failed->loaded() &&
-            $failed->count > 5 &&
-            (time() - $failed->time < 60));
+        $failed = ORM::factory('failed_auth')->where('name', '=', $name)->find();
+        return ($failed->loaded()
+                && $failed->count > 5
+                && (time() - $failed->time < 60));
     }
 
     public static function validate_too_many_failed_logins($name_input)
@@ -106,9 +93,7 @@ class auth_Core
      */
     public static function record_failed_attempt($name)
     {
-        $failed = ORM::factory('failed_auth')
-      ->where('name', '=', $name)
-      ->find();
+        $failed = ORM::factory('failed_auth')->where('name', '=', $name)->find();
         if (!$failed->loaded()) {
             $failed->name = $name;
         }
@@ -122,9 +107,7 @@ class auth_Core
      */
     public static function clear_failed_attempts($user)
     {
-        ORM::factory('failed_auth')
-      ->where('name', '=', $user->name)
-      ->delete_all();
+        ORM::factory('failed_auth')->where('name', '=', $user->name)->delete_all();
     }
 
     /**
@@ -138,10 +121,10 @@ class auth_Core
             access::forbidden();
         }
 
-        $session = Session::instance();
-        $last_active_auth = $session->get('active_auth_timestamp', 0);
+        $session                  = Session::instance();
+        $last_active_auth         = $session->get('active_auth_timestamp', 0);
         $last_admin_area_activity = $session->get('admin_area_activity_timestamp', 0);
-        $admin_area_timeout = module::get_var('gallery', 'admin_area_timeout');
+        $admin_area_timeout       = module::get_var('gallery', 'admin_area_timeout');
 
         if (max($last_active_auth, $last_admin_area_activity) + $admin_area_timeout < time()) {
             return true;
